Cambios en Test Actor-Category-Language

+++ Catalogo-Pruebas/src/main/java/com/example/domains/services/LanguageServiceImpl.java
@@ -18,8 +18,8 @@ import com.example.exceptions.InvalidDataException;
 import com.example.exceptions.NotFoundException;
 
 @Service
-public class LanguageServiceImpl implements LanguageService{
-	
+public class LanguageServiceImpl implements LanguageService {
+
 	@Autowired
 	LanguageRepository dao;
 
@@ -54,24 +54,24 @@ public class LanguageServiceImpl implements LanguageService{
 	}
 
 	@Override
-	public Optional<Language> getOne(Integer id) {
+	public Optional<Language> getOne(Integer id) throws Exception {
+		if (id < 0) {
+			throw new Exception("ERROR. La id no puede ser menor a 0");
+		}
 		return dao.findById(id);
 	}
 
 	@Override
 	public Language add(Language item) throws DuplicateKeyException, InvalidDataException {
-		if(item == null)
-		{
+		if (item == null) {
 			throw new InvalidDataException("[ERROR] El item no puede ser nulo");
 		}
-		
-		if(item.isInvalid())
-		{
+
+		if (item.isInvalid()) {
 			throw new InvalidDataException(item.getErrorsMessage());
 		}
-		
-		if(dao.existsById(item.getLanguageId()))
-		{
+
+		if (dao.existsById(item.getLanguageId())) {
 			throw new DuplicateKeyException(item.getErrorsMessage());
 		}
 		return dao.save(item);
@@ -79,18 +79,15 @@ public class LanguageServiceImpl implements LanguageService{
 
 	@Override
 	public Language modify(Language item) throws NotFoundException, InvalidDataException {
-		if(item == null)
-		{
+		if (item == null) {
 			throw new InvalidDataException("[ERROR] El item no puede ser nulo");
 		}
-		
-		if(item.isInvalid())
-		{
+
+		if (item.isInvalid()) {
 			throw new InvalidDataException(item.getErrorsMessage());
 		}
-		
-		if(!dao.existsById(item.getLanguageId()))
-		{
+
+		if (!dao.existsById(item.getLanguageId())) {
 			throw new NotFoundException();
 		}
 		return dao.save(item);
@@ -98,17 +95,23 @@ public class LanguageServiceImpl implements LanguageService{
 
 	@Override
 	public void delete(Language item) throws InvalidDataException {
-		if(item == null)
-		{
+		if (item == null) {
 			throw new InvalidDataException("[ERROR] El item no puede ser nulo");
 		}
-		
+
 		deleteById(item.getLanguageId());
 	}
 
 	@Override
-	public void deleteById(Integer id) {
-		dao.deleteById(id);
+	public void deleteById(Integer id) throws InvalidDataException {
+		try {
+			if (id < 0) {
+				throw new InvalidDataException("ERROR. La id no puede ser menor a 0");
+			}
+			dao.deleteById(id);
+		} catch (InvalidDataException e) {
+			throw new InvalidDataException(e.getMessage());
+		}
 	}
 
 	@Override
@@ -117,7 +120,4 @@ public class LanguageServiceImpl implements LanguageService{
 
 	}
 
-	
-	
-	
 }
